requirements.txt
cryptography>=41.0.1
pycryptodome>=3.18.0
pyca>=0.6.0
argon2-cffi>=21.3.0  # Optional, for Argon2 password hashing
